+++ android/src/main/kotlin/hamza/dali/flutter_osm_plugin/FlutterOsmView.kt
@@ -517,6 +517,9 @@ class FlutterOsmView(
                 "update#Marker" -> {
                     updateMarker(call, result)
                 }
+                "get#geopoints" -> {
+                    getGeoPoints(result)
+                }
                 else -> {
                     result.notImplemented()
                 }
@@ -526,6 +529,18 @@ class FlutterOsmView(
             Log.e(e.cause.toString(), "error osm plugin ${e.stackTraceToString()}")
             result.error("404", e.message, e.stackTraceToString())
         }
+    }
+
+    private fun getGeoPoints(result: MethodChannel.Result) {
+       val list = folderMarkers.items.filterIsInstance(Marker::class.java)
+        val geoPoints  = emptyList<HashMap<String,Double>>().toMutableList()
+        geoPoints.addAll(
+            list.map {
+                it.position.toHashMap()
+            }.toList()
+        )
+        result.success(geoPoints.toList())
+
     }
 
     private fun getUserLocation(result: MethodChannel.Result, callback: VoidCallback? = null) {
@@ -612,7 +627,6 @@ class FlutterOsmView(
                 }
                 val polyLine = Polyline(map!!)
                 polyLine.setPoints(lastRoad.roadPoints)
-                //customRoadMarkerIcon.p
                 flutterRoad = createRoad(
                     polyLine = polyLine,
                     colorRoad = lastRoad.roadColor,
@@ -1035,45 +1049,6 @@ class FlutterOsmView(
     }
 
 
-    private fun drawRoadManually(call: MethodCall, result: MethodChannel.Result) {
-        val args: HashMap<String, Any> = call.arguments as HashMap<String, Any>
-
-        val encodedWayPoints = (args["road"] as String)
-        val colorRoad = (args["roadColor"] as List<Int>)
-        val color = Color.rgb(colorRoad.first(), colorRoad.last(), colorRoad[1])
-        val widthRoad = (args["roadWidth"] as Double)
-        val zoomToRegion = args["zoomInto"] as Boolean
-        checkRoadFolderAboveUserOverlay()
-
-        folderRoad.items.clear()
-
-
-        val route = PolylineEncoder.decode(encodedWayPoints, 10, false)
-        val polyLine = Polyline(map!!)
-        polyLine.setPoints(route)
-        polyLine.outlinePaint.color = color
-        polyLine.outlinePaint.strokeWidth = widthRoad.toFloat()
-
-        folderRoad.items.add(polyLine)
-
-        mapSnapShot().cacheRoad(
-            RoadSnapShot(
-                roadPoints = route,
-                roadColor = color,
-                roadWith = widthRoad.toFloat(),
-                showIcons = false,
-            )
-        )
-        if (zoomToRegion) {
-            map!!.zoomToBoundingBox(
-                BoundingBox.fromGeoPoints(polyLine.actualPoints),
-                true,
-                64,
-            )
-        }
-        map!!.invalidate()
-        result.success(null)
-    }
 
     private fun trackUserLocation(result: MethodChannel.Result) {
         try {
@@ -1568,12 +1543,72 @@ class FlutterOsmView(
         }
     }
 
+    private fun drawRoadManually(call: MethodCall, result: MethodChannel.Result) {
+        val args: HashMap<String, Any> = call.arguments as HashMap<String, Any>
+
+        val encodedWayPoints = (args["road"] as String)
+        val colorRoad = (args["roadColor"] as List<Int>)
+        val color = Color.rgb(colorRoad.first(), colorRoad.last(), colorRoad[1])
+        val widthRoad = (args["roadWidth"] as Double)
+        val zoomToRegion = args["zoomInto"] as Boolean
+        val clearPreviousRoad = args["clearPreviousRoad"] as Boolean
+        val interestPointsEncoded = args["interestPoints"] as String?
+        val iconInterestPoints = args["iconInterestPoints"] as ByteArray?
+        checkRoadFolderAboveUserOverlay()
+        if (clearPreviousRoad) {
+            folderRoad.items.clear()
+        }
+        var bitmapIconInterestPoints: Bitmap? = null
+        if (iconInterestPoints != null) {
+            bitmapIconInterestPoints = getBitmap(bytes = iconInterestPoints!!)
+        }
+
+
+        val route = PolylineEncoder.decode(encodedWayPoints, 10, false)
+        val listInterestPoints = when (interestPointsEncoded != null) {
+            true -> PolylineEncoder.decode(interestPointsEncoded, 10, false)
+            false -> emptyList<GeoPoint>()
+        }
+
+        val polyLine = Polyline(map!!)
+        polyLine.setPoints(route)
+        polyLine.outlinePaint.color = color
+        polyLine.outlinePaint.strokeWidth = widthRoad.toFloat()
+        createRoad(
+            polyLine = polyLine,
+            colorRoad = color,
+            roadWidth = widthRoad.toFloat(),
+            showPoiMarker = listInterestPoints.isNotEmpty(),
+            listInterestPoints = listInterestPoints,
+            bitmapIcon = bitmapIconInterestPoints
+        )
+
+        mapSnapShot().cacheRoad(
+            RoadSnapShot(
+                roadPoints = route,
+                roadColor = color,
+                roadWith = widthRoad.toFloat(),
+                showIcons = false,
+            )
+        )
+        if (zoomToRegion) {
+            map!!.zoomToBoundingBox(
+                BoundingBox.fromGeoPoints(polyLine.actualPoints),
+                true,
+                64,
+            )
+        }
+        map!!.invalidate()
+        result.success(null)
+    }
+
     private fun createRoad(
         polyLine: Polyline,
         colorRoad: Int?,
         showPoiMarker: Boolean,
         listInterestPoints: List<GeoPoint>,
         roadWidth: Float,
+        bitmapIcon: Bitmap? = null,
     ): FlutterRoad {
         polyLine.setOnClickListener { _, _, eventPos ->
             methodChannel.invokeMethod("receiveSinglePress", eventPos?.toHashMap())
@@ -1582,19 +1617,38 @@ class FlutterOsmView(
         /// set polyline color
         polyLine.outlinePaint.color = colorRoad ?: Color.GREEN
 
-
+        val iconsRoads = customRoadMarkerIcon
+        when {
+            (iconsRoads.isEmpty() && bitmapIcon != null) -> {
+                iconsRoads[Constants.STARTPOSITIONROAD] = bitmapIcon
+                iconsRoads[Constants.MIDDLEPOSITIONROAD] = bitmapIcon
+                iconsRoads[Constants.ENDPOSITIONROAD] = bitmapIcon
+            }
+            iconsRoads.isNotEmpty() && bitmapIcon != null -> {
+                iconsRoads[Constants.MIDDLEPOSITIONROAD] = bitmapIcon
+                if (!iconsRoads.containsKey(Constants.STARTPOSITIONROAD)) {
+                    iconsRoads[Constants.STARTPOSITIONROAD] = bitmapIcon
+                }
+                if (!iconsRoads.containsKey(Constants.ENDPOSITIONROAD)) {
+                    iconsRoads[Constants.ENDPOSITIONROAD] = bitmapIcon
+                }
+            }
+        }
         val flutterRoad = FlutterRoad(
             context,
             map!!,
-            interestPoint = if (showPoiMarker) listInterestPoints else emptyList()
+            interestPoint = if (showPoiMarker) listInterestPoints else emptyList(),
+            showInterestPoints = showPoiMarker
         )
 
         flutterRoad.let { roadF ->
-            roadF.markersIcons = customRoadMarkerIcon
+            if (showPoiMarker) {
+                roadF.markersIcons = iconsRoads
+            }
             polyLine.outlinePaint.strokeWidth = roadWidth
 
             roadF.road = polyLine
-            if (showPoiMarker) {
+            /*if (showPoiMarker) {
                 // if (it.start != null)
                 folderRoad.items.add(roadF.start.apply {
                     this.visibilityInfoWindow(visibilityInfoWindow)
@@ -1604,9 +1658,9 @@ class FlutterOsmView(
                     this.visibilityInfoWindow(visibilityInfoWindow)
                 })
                 folderRoad.items.addAll(roadF.middlePoints)
-            }
+            }*/
 
-            folderRoad.items.add(roadF.road!!)
+            folderRoad.items.add(roadF)
         }
 
         return flutterRoad