+++ flutter_osm_interface/lib/src/channel/osm_method_channel.dart
@@ -494,10 +494,13 @@ class MethodChannelOSM extends MobileOSMPlatform {
   @override
   Future<void> drawRoadManually(
     int idOSM,
-    List<GeoPoint> road,
-    Color roadColor,
-    double width, {
+    List<GeoPoint> road, {
+    Color roadColor = Colors.green,
+    double width = 5.0,
     bool zoomInto = false,
+    bool deleteOldRoads = false,
+    GlobalKey? keyIconForInterestPoints,
+    List<GeoPoint> interestPoints = const [],
   }) async {
     final coordinates = road.map((e) => e.toListNum()).toList();
     final encodedCoordinates = encodePolyline(coordinates);
@@ -511,6 +514,19 @@ class MethodChannelOSM extends MobileOSMPlatform {
       data.addAll(roadColor.toMap("roadColor"));
     }
     data["zoomInto"] = zoomInto;
+    data["clearPreviousRoad"] = deleteOldRoads;
+    data["iconInterestPoints"] = null;
+    data["interestPoints"] = null;
+    if (interestPoints.isNotEmpty) {
+      data["interestPoints"] = await interestPoints.encodedToString();
+      if (keyIconForInterestPoints != null) {
+        try {
+          final Uint8List bytes = await _capturePng(keyIconForInterestPoints);
+          data["iconInterestPoints"] = Platform.isIOS ? bytes.convertToString() : bytes;
+        } catch (e) {}
+      }
+    }
+
     await _channels[idOSM]?.invokeMethod(
       "drawRoad#manually",
       data,
@@ -685,6 +701,12 @@ class MethodChannelOSM extends MobileOSMPlatform {
         result.map((e) => Map<String, dynamic>.from(e)).toList();
     return mapRoadInfo.map((e) => RoadInfo.fromMap(e)).toList();
   }
+
+  @override
+  Future<List<GeoPoint>> getGeoPointMarkers(int idOSM) async {
+    final list = await _channels[idOSM]!.invokeListMethod("get#geopoints");
+    return (list as List).map((e) => GeoPoint.fromMap(e)).toList();
+  }
 }
 
 extension config on MethodChannelOSM {
@@ -705,8 +727,10 @@ extension config on MethodChannelOSM {
     await _channels[idOSM]?.invokeMethod('config#Zoom', args);
   }
 
-  Future<void> initIosMap(int idOSM) async {
+  Future<void> initIosMap(int idOSM, GlobalKey key) async {
     await _channels[idOSM]?.invokeMethod("init#ios#map");
+    final icon = (await _capturePng(key)).convertToString();
+    await _channels[idOSM]?.invokeMethod("setDefaultIOSIcon", icon);
   }
 }