+++ lib/src/controller/osm/osm_controller.dart
@@ -65,7 +65,16 @@ class MobileOSMController extends IBaseOSMController {
 
     /// load config map scene for iOS
     if (Platform.isIOS) {
-      await (osmPlatform as MethodChannelOSM).initIosMap(_idMap);
+      _osmFlutterState.widget.dynamicMarkerWidgetNotifier.value = MarkerIcon(
+        icon: Icon(
+          Icons.location_on,
+          color: Colors.red,
+          size: 24,
+        ),
+      );
+      await Future.delayed(Duration(milliseconds: 300));
+      await (osmPlatform as MethodChannelOSM).initIosMap(_idMap, _osmFlutterState.dynamicMarkerKey);
+      _osmFlutterState.widget.dynamicMarkerWidgetNotifier.value = null;
     }
 
     _checkBoundingBox(box, initPosition);
@@ -260,8 +269,8 @@ class MobileOSMController extends IBaseOSMController {
       _idMap,
       [
         _osmFlutterState.startIconKey,
-        _osmFlutterState.endIconKey,
         _osmFlutterState.middleIconKey,
+        _osmFlutterState.endIconKey,
       ],
     );
   }
@@ -380,9 +389,11 @@ class MobileOSMController extends IBaseOSMController {
 
   /// activate current location position
   Future<void> currentLocation() async {
-    bool granted = await _osmFlutterState.requestPermission();
-    if (!granted) {
-      throw Exception("Location permission not granted");
+    if(Platform.isAndroid){
+      bool granted = await _osmFlutterState.requestPermission();
+      if (!granted) {
+        throw Exception("Location permission not granted");
+      }
     }
     // bool isEnabled = await _osmFlutterState.checkService();
     // if (!isEnabled) {
@@ -413,7 +424,7 @@ class MobileOSMController extends IBaseOSMController {
     MarkerIcon? markerIcon,
     double? angle,
   }) async {
-    if (markerIcon != null && (markerIcon.icon != null || markerIcon.assetMarker != null)) {
+    if (markerIcon != null) {
       _osmFlutterState.widget.dynamicMarkerWidgetNotifier.value =
           ((angle == null) || (angle == 0.0))
               ? markerIcon
@@ -421,7 +432,7 @@ class MobileOSMController extends IBaseOSMController {
                   angle: angle,
                   child: markerIcon,
                 );
-      int duration = markerIcon.icon != null || markerIcon.assetMarker != null ? 300 : 350;
+      int duration = 500;
       await Future.delayed(Duration(milliseconds: duration), () async {
         await osmPlatform.addMarker(
           _idMap,
@@ -515,11 +526,15 @@ class MobileOSMController extends IBaseOSMController {
   /// draw road
   ///  [path] : (list) path of the road
   Future<void> drawRoadManually(
-    List<GeoPoint> path,
-    Color roadColor,
-    double width, {
+    List<GeoPoint> path, {
+    Color roadColor = Colors.green,
+    double width = 5.0,
     bool zoomInto = false,
+    bool deleteOldRoads = false,
+    MarkerIcon? interestPointIcon,
+    List<GeoPoint> interestPoints = const [],
   }) async {
+    assert(width > 0.0);
     if (path.isEmpty) {
       throw Exception("you cannot make road with empty list of  geoPoint");
     }
@@ -528,12 +543,30 @@ class MobileOSMController extends IBaseOSMController {
         path.length < 3) {
       throw Exception("you cannot make line with same geoPoint");
     }
+    var icon = interestPointIcon;
+    if (Platform.isIOS && icon == null && interestPoints.isNotEmpty) {
+      icon = MarkerIcon(
+        icon: Icon(
+          Icons.location_on,
+          color: Colors.red,
+          size: 32,
+        ),
+      );
+    }
+    if (icon != null && interestPoints.isNotEmpty) {
+      _osmFlutterState.widget.dynamicMarkerWidgetNotifier.value = icon;
+      await Future.delayed(Duration(milliseconds: 350));
+    }
     await osmPlatform.drawRoadManually(
       _idMap,
       path,
-      roadColor,
-      width,
+      roadColor: roadColor,
+      width: width,
       zoomInto: zoomInto,
+      deleteOldRoads: deleteOldRoads,
+      interestPoints: interestPoints,
+      keyIconForInterestPoints:
+          interestPointIcon != null ? _osmFlutterState.dynamicMarkerKey : null,
     );
   }
 
@@ -692,6 +725,11 @@ class MobileOSMController extends IBaseOSMController {
       configs,
       commonRoadOption: commonRoadOption,
     );
+  }
+
+  @override
+  Future<List<GeoPoint>> geoPoints() async{
+    return await osmPlatform.getGeoPointMarkers(_idMap);
   }
 }