+++ lib/src/osm_flutter.dart
@@ -176,7 +176,7 @@ class _OSMFlutterState extends State<OSMFlutter> {
                         Container(
                           color: Colors.white,
                           child: buildWidget(
-                              controller: widget.controller as MapController,
+                              controller: widget.controller,
                               onGeoPointClicked: widget.onGeoPointClicked,
                               onLocationChanged: widget.onLocationChanged,
                               dynamicMarkerWidgetNotifier: dynamicMarkerWidgetNotifier,